validação dos campos do DTO

+++ src/main/java/br/com/marcus/ecommerce/services/ProductService.java
@@ -3,13 +3,17 @@ package br.com.marcus.ecommerce.services;
 import br.com.marcus.ecommerce.dto.ProductDTO;
 import br.com.marcus.ecommerce.entities.Product;
 import br.com.marcus.ecommerce.repositories.ProductRepository;
+import br.com.marcus.ecommerce.services.exeptions.DatabaseException;
 import br.com.marcus.ecommerce.services.exeptions.ResourceNotFoundExeption;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -39,6 +43,7 @@ public class ProductService {
     }
 
     public ProductDTO insert(ProductDTO dto) {
+        
         Product product = new Product();
         copyDtoToEntity(dto, product);
         productRepository.save(product);
@@ -47,10 +52,15 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product product = productRepository.getOne(id);
-        copyDtoToEntity(dto, product);
-        productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(dto, product);
+            productRepository.save(product);
+            return new ProductDTO(product);
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundExeption("Id para atualizar não encontrado");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product product) {
@@ -62,10 +72,16 @@ public class ProductService {
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundExeption("Id para deletar nao encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
-
 
 }
